Update route example
