+++ app/api/hello/route.ts
@@ -1,0 +1,3 @@
+export async function GET() {
+	return Response.json({ message: 'Hello World' });
+}
