--- components/HomePage/HomePage.stories.tsx
@@ -1,12 +1,0 @@
-import { HomePage } from '.';
-
-const meta = {
-	title: 'Pages/Home',
-	component: HomePage,
-};
-
-export default meta;
-
-export const Default = () => {
-	return <HomePage />;
-};
