--- components/HomePage/HomePage.tsx
@@ -1,26 +1,0 @@
-import { PageContent } from '@ag.ds-next/react/content';
-import { Prose } from '@ag.ds-next/react/prose';
-import { AppLayout } from '../AppLayout';
-
-export const HomePage = () => {
-	return (
-		<AppLayout>
-			<PageContent>
-				<Prose>
-					<h1>AG Design System Starter Kit</h1>
-					<p>
-						This is a simple starter kit built using the AG Design System,
-						NextJS and Typescript.
-					</p>
-					<p>
-						For examples of common user interfaces, please refer to the{' '}
-						<a href="https://design-system.agriculture.gov.au/templates">
-							design system templates.
-						</a>
-						.
-					</p>
-				</Prose>
-			</PageContent>
-		</AppLayout>
-	);
-};
