+++ components/HomePageContent/HomePageContent.stories.tsx
@@ -1,0 +1,12 @@
+import { HomePageContent } from './HomePageContent';
+
+const meta = {
+	title: 'Pages/Home',
+	component: HomePageContent,
+};
+
+export default meta;
+
+export const Default = () => {
+	return <HomePageContent />;
+};
